Removed casting of 0 to float64

+++ homework_1/Task-02/calculator/calculator.go
@@ -11,8 +11,8 @@ type TaxBracket struct {
 }
 
 func validateBracketsIncreasing(taxBrackets []TaxBracket) error {
-	lastThreshold := float64(0)
-	lastTaxRate := float64(0)
+	lastThreshold := 0.0
+	lastTaxRate := 0.0
 
 	for _, taxBracket := range taxBrackets {
 		threshold := taxBracket.Threshold
